Replaced initial actions and reducers Moved initial code into old/
