+++ src/main/webapp/front-end/src/index.jsx
@@ -4,22 +4,23 @@ import {Router, Route, hashHistory} from 'react-router';
 import {createStore, applyMiddleware, compose} from 'redux';
 import {Provider} from 'react-redux';
 import reducer from './reducer';
-import {setState, loadTriggerElementSubTypes, loadTriggerElementSubTypesSuccess, loadTriggerElementSubTypesError, cleanState} from './action_creators';
+import {SET_USER} from './actions';
 import promise from 'redux-promise';
 import App from './components/App';
 import Home from './components/Home';
 import Register from './components/Register';
 
+const initialState = {
+  userId: null
+}
+
 const createStoreWithMiddleware = applyMiddleware(
   promise
 )(createStore);
-const store = createStoreWithMiddleware(reducer, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
+const store = createStoreWithMiddleware(reducer, initialState, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
 
-store.dispatch(cleanState());
-
-store.dispatch(loadTriggerElementSubTypes()).then((response) => {
-               !response.error ? store.dispatch(loadTriggerElementSubTypesSuccess(response.payload.data)) : store.dispatch(loadTriggerElementSubTypesError(response.error));
-           });
+// Set fake user id to test redux
+store.dispatch({ type: SET_USER, userId: 1 });
 
 const routes = <Route component={App}>
   <Route path="/" component={Home} />
