+++ src/main/webapp/front-end/src/old/components_old/IntegerVariableForm.jsx
