+++ src/main/webapp/front-end/src/old/components_old/VariableForm.jsx
