+++ src/main/webapp/front-end/src/old/triggerModel.js
