+++ src/main/webapp/front-end/src/old/utilities.js
